chore: segregate purely mass models and hybrid models
