+++ src/gwkokab/models/hybrids/__init__.py
@@ -1,0 +1,6 @@
+# Copyright 2023 The GWKokab Authors
+# SPDX-License-Identifier: Apache-2.0
+
+
+from ._npowerlawmgaussian import NPowerlawMGaussian as NPowerlawMGaussian
+from ._powerlawpeak import PowerlawPeak as PowerlawPeak
