+++ src/gwkokab/models/hybrids/_npowerlawmgaussian.py
