+++ src/gwkokab/models/mass/_models.py
@@ -8,21 +8,15 @@ from jax import lax, numpy as jnp, random as jrd
 from jax.scipy import special
 from jax.scipy.stats import norm, uniform
 from jaxtyping import Array, ArrayLike
-from numpyro.distributions import constraints, Distribution, Independent
+from numpyro.distributions import constraints, Distribution
 from numpyro.distributions.util import promote_shapes, validate_sample
 
-from ..models.spin import BetaFromMeanVar, IndependentSpinOrientationGaussianIsotropic
-from ..models.transformations import PrimaryMassAndMassRatioToComponentMassesTransform
-from ..utils.kernel import log_planck_taper_window
-from .constraints import mass_ratio_mass_sandwich, mass_sandwich
-from .redshift import PowerlawRedshift
-from .utils import (
+from ...utils.kernel import log_planck_taper_window
+from ..constraints import mass_ratio_mass_sandwich, mass_sandwich
+from ..utils import (
     doubly_truncated_power_law_icdf,
     doubly_truncated_power_law_log_norm_constant,
     doubly_truncated_power_law_log_prob,
-    ExtendedSupportTransformedDistribution,
-    JointDistribution,
-    ScaledMixture,
 )
 
 
@@ -344,76 +338,3 @@ class SmoothedTwoComponentPrimaryMassRatio(Distribution):
         log_Z_q = jnp.where(_Z_q <= 0, 0.0, jnp.log(safe_Z_q))
         log_prob_q = self._log_prob_q_unnorm(value) - log_Z_q
         return log_prob_m1 + log_prob_q
-
-
-def PowerlawPeak(
-    use_spin: bool = False,
-    use_redshift: bool = False,
-    use_tilt: bool = False,
-    validate_args: Optional[bool] = None,
-    **params: Array,
-) -> ScaledMixture:
-    # NOTE: If you change something here, please also change in
-    # kokab/one_powerlaw_one_peak/ppd.py
-    smoothing_model = ExtendedSupportTransformedDistribution(
-        SmoothedTwoComponentPrimaryMassRatio(
-            alpha=params["alpha"],
-            beta=params["beta"],
-            delta=params["delta"],
-            lambda_peak=params["lambda_peak"],
-            loc=params["loc"],
-            mmax=params["mmax"],
-            mmin=params["mmin"],
-            scale=params["scale"],
-            validate_args=validate_args,
-        ),
-        transforms=PrimaryMassAndMassRatioToComponentMassesTransform(),
-        validate_args=validate_args,
-    )
-
-    component_distributions = [smoothing_model]
-
-    if use_spin:
-        chi_dist = Independent(
-            BetaFromMeanVar(
-                mean=jnp.stack([params["chi_mean"], params["chi_mean"]], axis=-1),
-                variance=jnp.stack(
-                    [params["chi_variance"], params["chi_variance"]], axis=-1
-                ),
-                validate_args=validate_args,
-            ),
-            reinterpreted_batch_ndims=1,
-            validate_args=validate_args,
-        )
-        component_distributions.append(chi_dist)
-
-    if use_tilt:
-        tilt_dist = IndependentSpinOrientationGaussianIsotropic(
-            zeta=params["cos_tilt_zeta"],
-            scale1=params["cos_tilt_scale"],
-            scale2=params["cos_tilt_scale"],
-            validate_args=validate_args,
-        )
-
-        component_distributions.append(tilt_dist)
-
-    if use_redshift:
-        z_max = params["z_max"]
-        kappa = params["kappa"]
-        powerlaw_z = PowerlawRedshift(
-            z_max=z_max, kappa=kappa, validate_args=validate_args
-        )
-
-        component_distributions.append(powerlaw_z)
-
-    if len(component_distributions) > 1:
-        component_distributions = [
-            JointDistribution(*component_distributions, validate_args=validate_args)
-        ]
-
-    return ScaledMixture(
-        log_scales=jnp.asarray([params["log_rate"]]),
-        component_distributions=component_distributions,
-        support=component_distributions[0]._support,  # type: ignore
-        validate_args=validate_args,
-    )
