--- src/gwkokab/models/npowerlawmgaussian/__init__.py
@@ -1,10 +1,0 @@
-# Copyright 2023 The GWKokab Authors
-# SPDX-License-Identifier: Apache-2.0
-
-
-from ._model import (
-    build_eccentricity_distributions as build_eccentricity_distributions,
-    build_spin_distributions as build_spin_distributions,
-    build_tilt_distributions as build_tilt_distributions,
-    NPowerlawMGaussian as NPowerlawMGaussian,
-)
